Update Updater.ps1
Now v1.2 as this is the second working rework of the updater.ps1 script. Bug fixes might still get pushed

Now outputs to a log file in the same folder it installs to

Fancier ANSI codes so the windows 11 users or windows terminal users can see nice looking text. Sucks 2 suck ig if you're on versions before that, sorry. I really wanted to be able to bold the text :3

Waits for 3 seconds at the end, so it doesn't flash and look scary like a powershell script just installed a virus on ur pc whenever u launch ur computer. Sits there for a little bit to let user click and see what it is, but still starts minimized, so it isn't in the way (but not completely hidden from user).